+++ src/ml_tooling/plots/viz/classification_viz.py
@@ -1,4 +1,4 @@
-from typing import Optional
+from typing import Optional, List
 
 from matplotlib import pyplot as plt
 
@@ -30,12 +30,14 @@ class ClassificationVisualize(BaseVisualize):
 
         normalized: bool
             Whether or not to normalize annotated class counts
+
         threshold: float
             Threshold to use for classification - defaults to 0.5
 
         Returns
         -------
-        matplotlib.Axes
+        plt.Axes
+            Returns a Confusion Matrix plot
         """
 
         with plt.style.context(MPL_STYLESHEET):
@@ -45,23 +47,34 @@ class ClassificationVisualize(BaseVisualize):
                 self._data.test_y, y_pred, normalized, title, **kwargs
             )
 
-    def roc_curve(self, **kwargs) -> plt.Axes:
+    def roc_curve(self, labels: List[str] = None, **kwargs) -> plt.Axes:
         """
         Visualize a ROC curve for a classification estimator.
         Estimator must implement a `predict_proba` method
         Any kwargs are passed onto matplotlib
 
+        Parameters
+        ----------
+        labels: List of str
+            Labels to use for the class names if multi-class
+
+        kwargs : optional
+            Keyword arguments to pass on to matplotlib
+
         Returns
         -------
-        matplotlib.Axes
+        plt.Axes
+            Returns a ROC AUC plot
         """
         if not hasattr(self._estimator, "predict_proba"):
             raise VizError("Model must provide a 'predict_proba' method")
 
         with plt.style.context(MPL_STYLESHEET):
             title = f"ROC AUC - {self._estimator_name}"
-            y_proba = self._estimator.predict_proba(self._data.test_x)[:, 1]
-            return plot_roc_auc(self._data.test_y, y_proba, title=title, **kwargs)
+            y_proba = self._estimator.predict_proba(self._data.test_x)
+            return plot_roc_auc(
+                self._data.test_y, y_proba, title=title, labels=labels, **kwargs
+            )
 
     def lift_curve(self, **kwargs) -> plt.Axes:
         """
@@ -69,16 +82,21 @@ class ClassificationVisualize(BaseVisualize):
         Estimator must implement a `predict_proba` method
         Any kwargs are passed onto matplotlib
 
+        Parameters
+        ----------
+        kwargs : optional
+            Keyword arguments to pass on to matplotlib
+
         Returns
         -------
-        matplotlib.Axes
+        plt.Axes
         """
         with plt.style.context(MPL_STYLESHEET):
             title = f"Lift Curve - {self._estimator_name}"
-            y_proba = self._estimator.predict_proba(self._data.test_x)[:, 1]
+            y_proba = self._estimator.predict_proba(self._data.test_x)
             return plot_lift_curve(self._data.test_y, y_proba, title=title, **kwargs)
 
-    def pr_curve(self, **kwargs) -> plt.Axes:
+    def precision_recall_curve(self, labels: List[str] = None, **kwargs) -> plt.Axes:
         """
         Visualize a Precision-Recall curve for a classification estimator.
         Estimator must implement a `predict_proba` method.
@@ -86,6 +104,9 @@ class ClassificationVisualize(BaseVisualize):
 
         Parameters
         ----------
+        labels: List of str
+            Labels to use for the class names if multi-class
+
         kwargs : optional
             Keyword arguments to pass on to matplotlib
 
@@ -100,5 +121,7 @@ class ClassificationVisualize(BaseVisualize):
 
         with plt.style.context(MPL_STYLESHEET):
             title = f"Precision-Recall - {self._estimator_name}"
-            y_proba = self._estimator.predict_proba(self._data.test_x)[:, 1]
-            return plot_pr_curve(self._data.test_y, y_proba, title=title, **kwargs)
+            y_proba = self._estimator.predict_proba(self._data.test_x)
+            return plot_pr_curve(
+                self._data.test_y, y_proba, title=title, labels=labels, **kwargs
+            )